Revert "Добавьте файлы проекта."
This reverts commit 68226f3250c9ff211aa5ed606d64b2106465ee2f.
